Orchestrated Observation prior to Nightfalls
This is how the orchestrators communicate directives to Baby and handler (who is always present watching) by sending zero-intent messages/oinks, the city and state is a coded instruction for the appearance/morphology of the penis to be seen later at a seemingly unrelated date. Switzerland is the equivalent of "go to the wire" in Prince's Cream.  That's exactly what happened.  More coming.